+++ models/contacts.json
@@ -58,5 +58,11 @@
     "name": "Aleks Word",
     "email": "Wonec.element@celeris.com",
     "phone": "(542) 341-1111"
+  },
+  {
+    "id": "510b330a-c57e-4f4a-ae4d-c8ebdc45396d",
+    "name": "Arnold Shwarts",
+    "email": "AS.element@celeris.org",
+    "phone": "38030589725"
   }
 ]
